[be] refactor and fix testing setup
