--- server/app/api/test.py
@@ -1,10 +1,0 @@
-from fastapi import APIRouter
-from fastapi.responses import JSONResponse
-
-
-router = APIRouter()
-
-
-@router.get("/")
-async def test_get() -> JSONResponse:
-    return JSONResponse({"message": "Hello from Picasso"})
