+++ server/app/core/__init__.py
@@ -1,2 +1,2 @@
+from .settings import settings
 from .logger import log
-from .config import settings
